new: direct sub component access can now be more explicit by prefixing with their format name.
Color instances would have to infer the target format based with the
name of the attribute. For instance ``Color(..).red`` would refer to
RGB format only because it is the only format being a ``namedtuple`` AND
having a component named 'red'.

This behavior is kept if it is not ambiguous, and it is now allowed to
be more specific (to avoid ambiguous cases) by prefixing by the format
name.

For instance, ``.red`` is equivalent to ``.rgb_red``.